Initial checkout refacto
