+++ _dev/js/theme/core/checkout/handler/editAddressHandler.js
@@ -1,0 +1,14 @@
+import prestashop from 'prestashop';
+
+const editAddressHandler = (event) => {
+  const {
+    addressesStep,
+    stepEdit,
+  } = prestashop.selectors.checkout;
+
+  event.stopPropagation();
+  document.querySelector(`${addressesStep} ${stepEdit}`)?.click();
+  prestashop.emit('editAddress');
+};
+
+export default editAddressHandler;
