+++ _dev/js/theme/core/checkout/utils/switchConfirmAddressesButtonState.js
@@ -1,0 +1,10 @@
+/**
+ * Enable/disable the continue address button
+ */
+const switchConfirmAddressesButtonState = (enable) => {
+  document.querySelectorAll('button[name=confirm-addresses]').forEach((button) => {
+    button.disabled = !enable;
+  });
+};
+
+export default switchConfirmAddressesButtonState;
